prep for more gpu tupes and more
21-install_nvidia_operator.sh
-fixed typo

31-install_ollama.sh
-prepared for additional gpu types
-added additional variables for (nvidia) GPU config 

34-install_cert-manager.sh
-enabled cert-manager to be installed in a namespace separate from suse-ai
-enabled check to see if cert-manager is already installed

35-install_open-webui_with_ollama.sh
-prepared for additional gpu types
-added additional variables for (nvidia) GPU config 

deploy_suse_ai.cfg
-rearranged config sections in the file
-changed the default SUSE AI namespace to be suse-private-ai and the cert-manager namespace to be cert-manager
-added configuration for deploying PyTorch
-added configuration for deploying suse-ai-deployer chart
-added additional variables for Ollama regarding (nvidia) GPUs

+++ 98-reset_cluster.sh
@@ -5,17 +5,41 @@ source deploy_suse_ai.cfg
 echo
 echo "Resetting the Cluster ..."
 echo
-echo "COMMAND: helm -n ${SUSE_AI_NAMESPACE} uninstall open-webui"
-helm -n ${SUSE_AI_NAMESPACE} uninstall open-webui
-echo
 
-echo "COMMAND: helm -n ${SUSE_AI_NAMESPACE} uninstall ollama"
-helm -n ${SUSE_AI_NAMESPACE} uninstall ollama
-echo
+if helm -n ${SUSE_AI_NAMESPACE} list | awk '{ print $1 }' | grep -q ${SUSE_AI_NAMESPACE}
+then
+  echo "COMMAND: helm -n ${SUSE_AI_NAMESPACE} uninstall ${SUSE_AI_NAMESPACE}"
+  helm -n ${SUSE_AI_NAMESPACE} uninstall ${SUSE_AI_NAMESPACE}
+  echo
+else
+  if helm -n ${SUSE_AI_NAMESPACE} | grep -q open-webui
+  then
+    echo "COMMAND: helm -n ${SUSE_AI_NAMESPACE} uninstall open-webui"
+    helm -n ${SUSE_AI_NAMESPACE} uninstall open-webui
+    echo
+  fi
 
-echo "COMMAND: helm -n ${SUSE_AI_NAMESPACE} uninstall milvus"
-helm -n ${SUSE_AI_NAMESPACE} uninstall milvus
-echo
+  if helm -n ${SUSE_AI_NAMESPACE} | grep -q ollama
+  then
+    echo "COMMAND: helm -n ${SUSE_AI_NAMESPACE} uninstall ollama"
+    helm -n ${SUSE_AI_NAMESPACE} uninstall ollama
+    echo
+  fi
+
+  if helm -n ${SUSE_AI_NAMESPACE} | grep -q milvus
+  then
+    echo "COMMAND: helm -n ${SUSE_AI_NAMESPACE} uninstall milvus"
+    helm -n ${SUSE_AI_NAMESPACE} uninstall milvus
+    echo
+  fi
+
+  if helm -n ${SUSE_AI_NAMESPACE} | grep -q pytorch
+  then
+    echo "COMMAND: helm -n ${SUSE_AI_NAMESPACE} uninstall pytorch"
+    helm -n ${SUSE_AI_NAMESPACE} uninstall pytorch
+    echo
+  fi
+fi
 
 echo "COMMAND: kubectl delete namespace ${SUSE_AI_NAMESPACE}"
 kubectl delete namespace ${SUSE_AI_NAMESPACE}
